+++ Makefile.am
@@ -17,7 +17,10 @@ ACLOCAL_AMFLAGS = -I m4
 AM_CPPFLAGS = -I$(top_srcdir)/include \
               -I$(top_srcdir)/src
 
-include_HEADERS=include/uv.h include/uv-errno.h include/uv-threadpool.h include/uv-version.h
+include_HEADERS=include/uv.h
+
+uvincludedir = $(includedir)/uv
+uvinclude_HEADERS=include/uv/errno.h include/uv/threadpool.h include/uv/version.h
 
 CLEANFILES =
 
@@ -42,7 +45,7 @@ endif
 
 if WINNT
 
-include_HEADERS += include/uv-win.h include/tree.h
+uvinclude_HEADERS += include/uv/win.h include/uv/tree.h
 AM_CPPFLAGS += -I$(top_srcdir)/src/win \
                -DWIN32_LEAN_AND_MEAN \
                -D_WIN32_WINNT=0x0600
@@ -82,7 +85,7 @@ libuv_la_SOURCES += src/win/async.c \
 
 else  # WINNT
 
-include_HEADERS += include/uv-unix.h
+uvinclude_HEADERS += include/uv/unix.h
 AM_CPPFLAGS += -I$(top_srcdir)/src/unix
 libuv_la_SOURCES += src/unix/async.c \
                    src/unix/atomic-ops.h \
@@ -333,13 +336,13 @@ libuv_la_CFLAGS += -D_ALL_SOURCE \
                    -D_LINUX_SOURCE_COMPAT \
                    -D_THREAD_SAFE \
                    -DHAVE_SYS_AHAFS_EVPRODS_H
-include_HEADERS += include/uv-aix.h
+uvinclude_HEADERS += include/uv/aix.h
 libuv_la_SOURCES += src/unix/aix.c src/unix/aix-common.c
 endif
 
 if ANDROID
-include_HEADERS += include/android-ifaddrs.h \
-                   include/pthread-barrier.h
+uvinclude_HEADERS += include/uv/android-ifaddrs.h \
+                   include/uv/pthread-barrier.h
 libuv_la_SOURCES += src/unix/android-ifaddrs.c \
                     src/unix/pthread-fixes.c
 endif
@@ -359,8 +362,8 @@ libuv_la_SOURCES += src/unix/cygwin.c \
 endif
 
 if DARWIN
-include_HEADERS += include/uv-darwin.h \
-                   include/pthread-barrier.h
+uvinclude_HEADERS += include/uv/darwin.h \
+                   include/uv/pthread-barrier.h
 libuv_la_CFLAGS += -D_DARWIN_USE_64_BIT_INODE=1
 libuv_la_CFLAGS += -D_DARWIN_UNLIMITED_SELECT=1
 libuv_la_SOURCES += src/unix/bsd-ifaddrs.c \
@@ -373,7 +376,7 @@ test_run_tests_LDFLAGS += -lutil
 endif
 
 if DRAGONFLY
-include_HEADERS += include/uv-bsd.h
+uvinclude_HEADERS += include/uv/bsd.h
 libuv_la_SOURCES += src/unix/bsd-ifaddrs.c \
                     src/unix/freebsd.c \
                     src/unix/kqueue.c \
@@ -382,7 +385,7 @@ test_run_tests_LDFLAGS += -lutil
 endif
 
 if FREEBSD
-include_HEADERS += include/uv-bsd.h
+uvinclude_HEADERS += include/uv/bsd.h
 libuv_la_SOURCES += src/unix/bsd-ifaddrs.c \
                     src/unix/freebsd.c \
                     src/unix/kqueue.c \
@@ -391,7 +394,7 @@ test_run_tests_LDFLAGS += -lutil
 endif
 
 if LINUX
-include_HEADERS += include/uv-linux.h
+uvinclude_HEADERS += include/uv/linux.h
 libuv_la_CFLAGS += -D_GNU_SOURCE
 libuv_la_SOURCES += src/unix/linux-core.c \
                     src/unix/linux-inotify.c \
@@ -418,7 +421,7 @@ libuv_la_SOURCES += src/unix/cygwin.c \
 endif
 
 if NETBSD
-include_HEADERS += include/uv-bsd.h
+uvinclude_HEADERS += include/uv/bsd.h
 libuv_la_SOURCES += src/unix/bsd-ifaddrs.c \
                     src/unix/kqueue.c \
                     src/unix/netbsd.c \
@@ -427,7 +430,7 @@ test_run_tests_LDFLAGS += -lutil
 endif
 
 if OPENBSD
-include_HEADERS += include/uv-bsd.h
+uvinclude_HEADERS += include/uv/bsd.h
 libuv_la_SOURCES += src/unix/bsd-ifaddrs.c \
                     src/unix/kqueue.c \
                     src/unix/openbsd.c \
@@ -436,14 +439,14 @@ test_run_tests_LDFLAGS += -lutil
 endif
 
 if SUNOS
-include_HEADERS += include/uv-sunos.h
+uvinclude_HEADERS += include/uv/sunos.h
 libuv_la_CFLAGS += -D__EXTENSIONS__ -D_XOPEN_SOURCE=500
 libuv_la_SOURCES += src/unix/no-proctitle.c \
                     src/unix/sunos.c
 endif
 
 if OS390
-include_HEADERS += include/pthread-barrier.h
+uvinclude_HEADERS += include/uv/pthread-barrier.h
 libuv_la_CFLAGS += -D_UNIX03_THREADS \
                    -D_UNIX03_SOURCE \
                    -D_OPEN_SYS_IF_EXT=1 \
