+++ include/uv/unix.h
@@ -42,32 +42,32 @@
 #include <pthread.h>
 #include <signal.h>
 
-#include "uv-threadpool.h"
+#include "uv/threadpool.h"
 
 #if defined(__linux__)
-# include "uv-linux.h"
+# include "uv/linux.h"
 #elif defined (__MVS__)
-# include "uv-os390.h"
+# include "uv/os390.h"
 #elif defined(__PASE__)
-# include "uv-posix.h"
+# include "uv/posix.h"
 #elif defined(_AIX)
-# include "uv-aix.h"
+# include "uv/aix.h"
 #elif defined(__sun)
-# include "uv-sunos.h"
+# include "uv/sunos.h"
 #elif defined(__APPLE__)
-# include "uv-darwin.h"
+# include "uv/darwin.h"
 #elif defined(__DragonFly__)       || \
       defined(__FreeBSD__)         || \
       defined(__FreeBSD_kernel__)  || \
       defined(__OpenBSD__)         || \
       defined(__NetBSD__)
-# include "uv-bsd.h"
+# include "uv/bsd.h"
 #elif defined(__CYGWIN__) || defined(__MSYS__)
-# include "uv-posix.h"
+# include "uv/posix.h"
 #endif
 
 #ifndef PTHREAD_BARRIER_SERIAL_THREAD
-# include "pthread-barrier.h"
+# include "uv/pthread-barrier.h"
 #endif
 
 #ifndef NI_MAXHOST
