+++ uv.gyp
@@ -64,10 +64,10 @@
       'sources': [
         'common.gypi',
         'include/uv.h',
-        'include/tree.h',
-        'include/uv-errno.h',
-        'include/uv-threadpool.h',
-        'include/uv-version.h',
+        'include/uv/tree.h',
+        'include/uv/errno.h',
+        'include/uv/threadpool.h',
+        'include/uv/version.h',
         'src/fs-poll.c',
         'src/heap-inl.h',
         'src/inet.c',
@@ -95,7 +95,7 @@
             '_GNU_SOURCE',
           ],
           'sources': [
-            'include/uv-win.h',
+            'include/uv/win.h',
             'src/win/async.c',
             'src/win/atomicops-inl.h',
             'src/win/core.c',
@@ -144,12 +144,12 @@
           },
         }, { # Not Windows i.e. POSIX
           'sources': [
-            'include/uv-unix.h',
-            'include/uv-linux.h',
-            'include/uv-sunos.h',
-            'include/uv-darwin.h',
-            'include/uv-bsd.h',
-            'include/uv-aix.h',
+            'include/uv/unix.h',
+            'include/uv/linux.h',
+            'include/uv/sunos.h',
+            'include/uv/darwin.h',
+            'include/uv/bsd.h',
+            'include/uv/aix.h',
             'src/unix/async.c',
             'src/unix/atomic-ops.h',
             'src/unix/core.c',
@@ -199,7 +199,7 @@
             ['uv_library=="shared_library" and OS!="mac" and OS!="zos"', {
               # This will cause gyp to set soname
               # Must correspond with UV_VERSION_MAJOR
-              # in include/uv-version.h
+              # in include/uv/version.h
               'product_extension': 'so.1',
             }],
           ],
